🧩 refactor music cog and modularize all classes
