+++ classes/__init__.py
@@ -1,0 +1,3 @@
+
+from .Song import *
+from .GuildState import *
