+++ cogs/music.py
@@ -4,10 +4,6 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 import asyncio
-import yt_dlp
-import functools
-from enum import Enum
-import math
 import logging
 import os
 import random
@@ -15,572 +11,15 @@ import aiohttp
 import re
 from typing import Union, Optional
 import google.generativeai as genai
+from classes import Song, GuildState
+from views import SearchView
 
 # === CONSTANTS & HELPERS ===
 log = logging.getLogger(__name__)
 AnyContext = Union[commands.Context, discord.Interaction]
-YTDL_SEARCH_OPTIONS = {
-    "format": "bestaudio/best",
-    "noplaylist": True,
-    "nocheckcertificate": True,
-    "ignoreerrors": False,
-    "logtostderr": False,
-    "quiet": True,
-    "no_warnings": True,
-    "default_search": "ytsearch7",
-    "source_address": "0.0.0.0",
-    "extract_flat": "search",
-}
-YTDL_DOWNLOAD_OPTIONS = {
-    "format": "bestaudio[ext=m4a]/bestaudio/best",
-    "outtmpl": "cache/%(id)s.%(ext)s",
-    "restrictfilenames": True,
-    "noplaylist": True,
-    "nocheckcertificate": True,
-    "ignoreerrors": False,
-    "logtostderr": False,
-    "quiet": True,
-    "no_warnings": True,
-    "source_address": "0.0.0.0",
-    "cachedir": False,
-}
-FFMPEG_OPTIONS = {"before_options": "", "options": "-vn"}
-
-
-class LoopMode(Enum):
-    OFF = 0
-    SONG = 1
-    QUEUE = 2
-
-
-# === DATA CLASSES ===
-class Song:
-    """Đại diện cho một bài hát."""
-
-    def __init__(self, data, requester: discord.Member | discord.User):
-        self.requester = requester
-        self.data = data
-        self.url = data.get("webpage_url") or data.get("url")
-        self.title = data.get("title")
-        self.thumbnail = data.get("thumbnail")
-        self.duration = data.get("duration")
-        self.uploader = data.get("uploader")
-        self.filepath = None
-        self.id = data.get("id")
-
-    def format_duration(self):
-        if self.duration is None:
-            return "N/A"
-        m, s = divmod(self.duration, 60)
-        h, m = divmod(m, 60)
-        return (
-            f"{int(h):02d}:{int(m):02d}:{int(s):02d}"
-            if h > 0
-            else f"{int(m):02d}:{int(s):02d}"
-        )
-
-    def cleanup(self):
-        if self.filepath and os.path.exists(self.filepath):
-            try:
-                os.remove(self.filepath)
-                log.info(f"Đã xóa file cache: {self.filepath}")
-            except OSError as e:
-                log.error(f"Lỗi khi xóa file cache {self.filepath}: {e}")
-
-    @classmethod
-    async def search_only(cls, query: str, requester: discord.Member | discord.User):
-        loop = asyncio.get_running_loop()
-        partial = functools.partial(
-            yt_dlp.YoutubeDL(YTDL_SEARCH_OPTIONS).extract_info, query, download=False
-        )
-        try:
-            data = await loop.run_in_executor(None, partial)
-            if not data or "entries" not in data or not data["entries"]:
-                return []
-            return [cls(entry, requester) for entry in data["entries"]]
-        except Exception as e:
-            log.error(f"Lỗi yt-dlp khi TÌM KIẾM '{query}': {e}", exc_info=True)
-            return []
-
-    @classmethod
-    async def from_url_and_download(
-        cls, url: str, requester: discord.Member | discord.User
-    ):
-        loop = asyncio.get_running_loop()
-        ytdl = yt_dlp.YoutubeDL(YTDL_DOWNLOAD_OPTIONS)
-        partial = functools.partial(ytdl.extract_info, url, download=True)
-        try:
-            data = await loop.run_in_executor(None, partial)
-            if not data:
-                return None
-            if "entries" in data:
-                data = data["entries"][0]
-            song = cls(data, requester)
-            song.filepath = ytdl.prepare_filename(data)
-            return song
-        except Exception as e:
-            log.error(f"Lỗi yt-dlp khi TẢI VỀ '{url}': {e}", exc_info=True)
-            return None
-
-
-class SearchView(discord.ui.View):
-    """Giao diện cho kết quả tìm kiếm."""
-
-    def __init__(self, *, music_cog, ctx: AnyContext, results: list[Song]):
-        super().__init__(timeout=180.0)
-        self.music_cog = music_cog
-        self.ctx = ctx
-        self.requester = ctx.author if isinstance(ctx, commands.Context) else ctx.user
-        self.results = results
-        self.current_page = 1
-        self.songs_per_page = 5
-        self.total_pages = math.ceil(len(self.results) / self.songs_per_page)
-        self.message = None
-        self.update_components()
-
-    async def on_timeout(self):
-        if self.message:
-            try:
-                await self.message.edit(
-                    content="Hết thời gian tìm kiếm.", embed=None, view=None
-                )
-            except discord.NotFound:
-                pass
-        self.stop()
-
-    async def start(self):
-        embed = self.create_page_embed()
-        if isinstance(self.ctx, discord.Interaction):
-            if self.ctx.response.is_done():
-                self.message = await self.ctx.followup.send(
-                    embed=embed, view=self, ephemeral=True
-                )
-            else:
-                await self.ctx.response.send_message(
-                    embed=embed, view=self, ephemeral=True
-                )
-                self.message = await self.ctx.original_response()
-        else:
-            self.message = await self.ctx.send(embed=embed, view=self)
-
-    def update_components(self):
-        self.prev_page_button.disabled = self.current_page == 1
-        self.next_page_button.disabled = self.current_page >= self.total_pages
-        self.clear_items()
-        self.add_item(self.create_select_menu())
-        self.add_item(self.prev_page_button)
-        self.add_item(self.next_page_button)
-        self.add_item(self.cancel_button)
-
-    def create_page_embed(self) -> discord.Embed:
-        start_index = (self.current_page - 1) * self.songs_per_page
-        end_index = start_index + self.songs_per_page
-        page_results = self.results[start_index:end_index]
-        description = "".join(
-            f"`{i+1}.` [{s.title}]({s.url})\n`{s.uploader or 'N/A'} - {s.format_duration()}`\n\n"
-            for i, s in enumerate(page_results, start=start_index)
-        )
-        embed = discord.Embed(
-            title=f"🔎 Kết quả tìm kiếm (Trang {self.current_page}/{self.total_pages})",
-            description=description,
-            color=discord.Color.blue(),
-        )
-        embed.set_footer(
-            text=f"Yêu cầu bởi {self.requester.display_name}",
-            icon_url=self.requester.display_avatar.url,
-        )
-        return embed
-
-    def create_select_menu(self) -> discord.ui.Select:
-        start_index = (self.current_page - 1) * self.songs_per_page
-        end_index = start_index + self.songs_per_page
-        options = [
-            discord.SelectOption(label=f"{i+1}. {s.title[:80]}", value=str(i))
-            for i, s in enumerate(
-                self.results[start_index:end_index], start=start_index
-            )
-        ]
-        select = discord.ui.Select(
-            placeholder="Chọn một bài hát để thêm...",
-            options=options,
-            custom_id="search_select_menu",
-        )
-        select.callback = self.select_callback
-        return select
-
-    async def select_callback(self, interaction: discord.Interaction):
-        if interaction.user.id != self.requester.id:
-            return await interaction.response.send_message(
-                "Bạn không phải người yêu cầu!", ephemeral=True
-            )
-        await interaction.response.defer()
-        await self.message.edit(
-            content="⏳ Đang tải bài hát bạn chọn...", embed=None, view=None
-        )
-        selected_song = await Song.from_url_and_download(
-            self.results[int(interaction.data["values"][0])].url, self.requester
-        )
-        if selected_song:
-            state = self.music_cog.get_guild_state(interaction.guild_id)
-            await state.queue.put(selected_song)
-            if state.player_task is None or state.player_task.done():
-                state.player_task = asyncio.create_task(state.player_loop())
-            await self.message.edit(
-                content=f"✅ Đã thêm **{selected_song.title}** vào hàng đợi."
-            )
-        else:
-            await self.message.edit(
-                content=f"❌ Rất tiếc, đã có lỗi khi tải về bài hát này."
-            )
-        self.stop()
-
-    @discord.ui.button(label="Trước", style=discord.ButtonStyle.secondary, emoji="⬅️")
-    async def prev_page_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
-    ):
-        if interaction.user.id != self.requester.id:
-            return await interaction.response.send_message(
-                "Bạn không phải người yêu cầu!", ephemeral=True
-            )
-        self.current_page -= 1
-        self.update_components()
-        await interaction.response.edit_message(
-            embed=self.create_page_embed(), view=self
-        )
-
-    @discord.ui.button(label="Sau", style=discord.ButtonStyle.secondary, emoji="➡️")
-    async def next_page_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
-    ):
-        if interaction.user.id != self.requester.id:
-            return await interaction.response.send_message(
-                "Bạn không phải người yêu cầu!", ephemeral=True
-            )
-        self.current_page += 1
-        self.update_components()
-        await interaction.response.edit_message(
-            embed=self.create_page_embed(), view=self
-        )
-
-    @discord.ui.button(label="Hủy", style=discord.ButtonStyle.danger, emoji="⏹️")
-    async def cancel_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
-    ):
-        if interaction.user.id != self.requester.id:
-            return await interaction.response.send_message(
-                "Bạn không phải người yêu cầu!", ephemeral=True
-            )
-        await self.message.edit(content="Đã hủy tìm kiếm.", embed=None, view=None)
-        self.stop()
-
-
-class GuildState:
-    """Quản lý trạng thái của từng server."""
-
-    def __init__(self, bot: commands.Bot, guild_id: int):
-        self.bot = bot
-        self.guild_id = guild_id
-        self.queue = asyncio.Queue[Song]()
-        self.voice_client: discord.VoiceClient | None = None
-        self.now_playing_message: discord.Message | None = None
-        self.current_song: Song | None = None
-        self.loop_mode = LoopMode.OFF
-        self.player_task: asyncio.Task | None = None
-        self.last_ctx: AnyContext | None = None
-        self.song_finished_event = asyncio.Event()
-        self.volume = 0.5
-        self.is_seeking = False
-
-    async def player_loop(self):
-        await self.bot.wait_until_ready()
-        while True:
-            self.song_finished_event.clear()
-
-            # Xử lý bài hát vừa phát xong (nếu có)
-            previous_song = self.current_song
-            if previous_song:
-                if self.loop_mode == LoopMode.QUEUE:
-                    await self.queue.put(previous_song)
-                elif self.loop_mode != LoopMode.SONG:
-                    previous_song.cleanup()
-
-            # Lấy bài hát tiếp theo
-            try:
-                # Nếu không lặp lại bài hát, lấy bài mới từ hàng đợi
-                if self.loop_mode != LoopMode.SONG:
-                    self.current_song = await asyncio.wait_for(
-                        self.queue.get(), timeout=300
-                    )
-                # Nếu lặp lại, self.current_song vẫn giữ nguyên
-            except asyncio.TimeoutError:
-                log.info(
-                    f"Guild {self.guild_id} không hoạt động trong 5 phút, bắt đầu dọn dẹp."
-                )
-                if self.last_ctx and self.last_ctx.channel:
-                    try:
-                        await self.last_ctx.channel.send(
-                            "😴 Đã tự động ngắt kết nối do không hoạt động."
-                        )
-                    except discord.Forbidden:
-                        pass
-                return await self.cleanup()
-
-            # Phát bài hát mới
-            try:
-                log.info(
-                    f"Guild {self.guild_id}: Lấy bài hát '{self.current_song.title}' từ hàng đợi."
-                )
-                await self.update_now_playing_message(new_song=True)
-                source = discord.PCMVolumeTransformer(
-                    discord.FFmpegPCMAudio(
-                        self.current_song.filepath, **FFMPEG_OPTIONS
-                    ),
-                    volume=self.volume,
-                )
-                self.voice_client.play(
-                    source,
-                    after=lambda e: self.bot.loop.call_soon_threadsafe(
-                        self.song_finished_event.set
-                    ),
-                )
-                await self.song_finished_event.wait()
-
-                # Bỏ qua logic sau khi phát xong nếu đang seek
-                if self.is_seeking:
-                    self.is_seeking = False
-                    continue
-
-                log.info(
-                    f"Guild {self.guild_id}: Sự kiện kết thúc bài hát '{self.current_song.title}' được kích hoạt."
-                )
-            except Exception as e:
-                log.error(
-                    f"Lỗi nghiêm trọng trong player loop của guild {self.guild_id}:",
-                    exc_info=e,
-                )
-                if self.last_ctx and self.last_ctx.channel:
-                    try:
-                        await self.last_ctx.channel.send(
-                            f"🤖 Gặp lỗi nghiêm trọng, Miku cần khởi động lại trình phát nhạc. Lỗi: `{e}`"
-                        )
-                    except discord.Forbidden:
-                        pass
-                return await self.cleanup()
-
-            # Kiểm tra nếu hàng đợi trống sau khi bài hát kết thúc
-            if self.queue.empty() and self.loop_mode == LoopMode.OFF:
-                log.info(f"Guild {self.guild_id}: Hàng đợi đã hết.")
-                if self.last_ctx and self.last_ctx.channel:
-                    try:
-                        await self.last_ctx.channel.send(
-                            "🎶 Hàng đợi đã kết thúc! Miku đi nghỉ đây (´｡• ᵕ •｡`) ♡"
-                        )
-                    except discord.Forbidden:
-                        pass
-                return await self.cleanup()
-
-    async def update_now_playing_message(self, new_song=False):
-        if not self.last_ctx:
-            return
-        if not self.current_song and self.now_playing_message:
-            try:
-                await self.now_playing_message.delete()
-            except discord.NotFound:
-                pass
-            self.now_playing_message = None
-            return
-        if not self.current_song:
-            return
-        embed = self.create_now_playing_embed()
-        view = self.create_control_view()
-        if new_song and self.now_playing_message:
-            try:
-                await self.now_playing_message.delete()
-            except discord.NotFound:
-                pass
-            self.now_playing_message = None
-        if self.now_playing_message:
-            try:
-                await self.now_playing_message.edit(embed=embed, view=view)
-                return
-            except discord.NotFound:
-                self.now_playing_message = None
-        if not self.now_playing_message:
-            try:
-                self.now_playing_message = await self.last_ctx.channel.send(
-                    embed=embed, view=view
-                )
-            except (discord.Forbidden, discord.HTTPException) as e:
-                log.warning(f"Không thể gửi/cập nhật tin nhắn Now Playing: {e}")
-                self.now_playing_message = None
-
-    def create_now_playing_embed(self) -> discord.Embed:
-        song = self.current_song
-        embed = discord.Embed(title=song.title, url=song.url, color=0x39D0D6)
-        embed.set_author(
-            name=f"Đang phát 🎵 (Âm lượng: {int(self.volume*100)}%)",
-            icon_url=self.bot.user.display_avatar.url,
-        )
-        embed.set_thumbnail(url=song.thumbnail)
-        embed.add_field(name="Nghệ sĩ", value=song.uploader or "N/A", inline=True)
-        embed.add_field(name="Thời lượng", value=song.format_duration(), inline=True)
-        embed.add_field(name="Yêu cầu bởi", value=song.requester.mention, inline=True)
-        loop_status = {
-            LoopMode.OFF: "Tắt",
-            LoopMode.SONG: "🔁 Bài hát",
-            LoopMode.QUEUE: "🔁 Hàng đợi",
-        }
-        next_song_title = (
-            "Không có"
-            if self.queue.empty()
-            else self.queue._queue[0].title[:50] + "..."
-        )
-        total_songs = self.queue.qsize() + (1 if self.current_song else 0)
-        embed.set_footer(
-            text=f"Tiếp theo: {next_song_title} | Lặp: {loop_status[self.loop_mode]} | Tổng cộng: {total_songs} bài"
-        )
-        return embed
-
-    def create_control_view(self) -> discord.ui.View:
-        view = discord.ui.View(timeout=None)
-        pause_resume_btn = discord.ui.Button(
-            emoji="⏯️",
-            style=discord.ButtonStyle.secondary,
-            custom_id=f"ctrl_pause_{self.guild_id}",
-        )
-        skip_btn = discord.ui.Button(
-            emoji="⏭️",
-            style=discord.ButtonStyle.secondary,
-            custom_id=f"ctrl_skip_{self.guild_id}",
-        )
-        stop_btn = discord.ui.Button(
-            emoji="⏹️",
-            style=discord.ButtonStyle.danger,
-            custom_id=f"ctrl_stop_{self.guild_id}",
-        )
-        loop_btn = discord.ui.Button(
-            emoji="🔁",
-            style=discord.ButtonStyle.secondary,
-            custom_id=f"ctrl_loop_{self.guild_id}",
-        )
-        queue_btn = discord.ui.Button(
-            label="Hàng đợi",
-            emoji="📜",
-            style=discord.ButtonStyle.primary,
-            custom_id=f"ctrl_queue_{self.guild_id}",
-        )
-        pause_resume_btn.callback = self.pause_resume_callback
-        skip_btn.callback = self.skip_callback
-        stop_btn.callback = self.stop_callback
-        loop_btn.callback = self.loop_callback
-        queue_btn.callback = self.queue_callback
-        view.add_item(pause_resume_btn)
-        view.add_item(skip_btn)
-        view.add_item(stop_btn)
-        view.add_item(loop_btn)
-        view.add_item(queue_btn)
-        return view
-
-    async def pause_resume_callback(self, interaction: discord.Interaction):
-        if self.voice_client.is_paused():
-            self.voice_client.resume()
-            await interaction.response.send_message(
-                "▶️ Đã tiếp tục phát.", ephemeral=True
-            )
-        else:
-            self.voice_client.pause()
-            await interaction.response.send_message("⏸️ Đã tạm dừng.", ephemeral=True)
-
-    async def skip_callback(self, interaction: discord.Interaction):
-        if self.voice_client and (
-            self.voice_client.is_playing() or self.voice_client.is_paused()
-        ):
-            self.voice_client.stop()
-            await interaction.response.send_message("⏭️ Đã chuyển bài.", ephemeral=True)
-        else:
-            await interaction.response.send_message(
-                "Không có bài nào đang phát để chuyển.", ephemeral=True
-            )
-
-    async def stop_callback(self, interaction: discord.Interaction):
-        await interaction.response.send_message(
-            "⏹️ Đang dừng phát nhạc và dọn dẹp hàng đợi...", ephemeral=True
-        )
-        await self.cleanup()
-
-    async def loop_callback(self, interaction: discord.Interaction):
-        self.loop_mode = LoopMode((self.loop_mode.value + 1) % 3)
-        log.info(f"Guild {self.guild_id} đã đổi chế độ lặp thành {self.loop_mode.name}")
-        mode_text = {
-            LoopMode.OFF: "Tắt lặp.",
-            LoopMode.SONG: "🔁 Lặp lại bài hát hiện tại.",
-            LoopMode.QUEUE: "🔁 Lặp lại toàn bộ hàng đợi.",
-        }
-        await interaction.response.send_message(
-            mode_text[self.loop_mode], ephemeral=True
-        )
-        await self.update_now_playing_message()
-
-    async def queue_callback(self, interaction: discord.Interaction):
-        embed = self._create_queue_embed()
-        if not embed:
-            return await interaction.response.send_message(
-                "Hàng đợi trống!", ephemeral=True
-            )
-        await interaction.response.send_message(embed=embed, ephemeral=True)
-
-    def _create_queue_embed(self) -> discord.Embed | None:
-        if self.queue.empty() and not self.current_song:
-            return None
-        embed = discord.Embed(title="📜 Hàng đợi bài hát", color=discord.Color.gold())
-        if self.current_song:
-            embed.add_field(
-                name="▶️ Đang phát",
-                value=f"[{self.current_song.title}]({self.current_song.url}) - Y/c bởi {self.current_song.requester.mention}",
-                inline=False,
-            )
-        queue_list = list(self.queue._queue)
-        if queue_list:
-            queue_text = "\n".join(
-                [
-                    f"`{i+1}.` [{song.title}]({song.url})"
-                    for i, song in enumerate(queue_list[:10])
-                ]
-            )
-            if len(queue_list) > 10:
-                queue_text += f"\n... và {len(queue_list) - 10} bài hát khác."
-            embed.add_field(name="🎶 Tiếp theo", value=queue_text, inline=False)
-        embed.set_footer(
-            text=f"Tổng cộng: {len(queue_list) + (1 if self.current_song else 0)} bài hát"
-        )
-        return embed
-
-    async def cleanup(self):
-        log.info(f"Bắt đầu cleanup cho guild {self.guild_id}")
-        self.bot.dispatch("session_end", self.guild_id)
-        if self.player_task:
-            self.player_task.cancel()
-        if self.current_song:
-            self.current_song.cleanup()
-            self.current_song = None
-        while not self.queue.empty():
-            try:
-                song = self.queue.get_nowait()
-                song.cleanup()
-            except asyncio.QueueEmpty:
-                break
-        if self.voice_client:
-            await self.voice_client.disconnect(force=True)
-            log.info(f"Đã ngắt kết nối voice client khỏi guild {self.guild_id}")
-        if self.now_playing_message:
-            try:
-                await self.now_playing_message.delete()
-            except discord.NotFound:
-                pass
-
 
 # === COG: MAIN ===
-class MainCog(commands.Cog, name="Miku"):
+class MusicCog(commands.Cog, name="Miku"):
     """Chứa toàn bộ lệnh và logic chính của bot."""
 
     music_group = app_commands.Group(
@@ -708,6 +147,7 @@ class MainCog(commands.Cog, name="Miku"):
         state = self.get_guild_state(ctx.guild.id)
         state.last_ctx = ctx
         author = ctx.author if isinstance(ctx, commands.Context) else ctx.user
+
         if not author.voice or not author.voice.channel:
             return await self._send_response(
                 ctx, "Bạn phải ở trong một kênh thoại để dùng lệnh này!", ephemeral=True
@@ -726,6 +166,7 @@ class MainCog(commands.Cog, name="Miku"):
                     ctx, "Không có nhạc nào đang phát hoặc tạm dừng.", ephemeral=True
                 )
             return
+        
         if isinstance(ctx, discord.Interaction):
             await ctx.response.defer(ephemeral=False)
         else:
@@ -738,13 +179,16 @@ class MainCog(commands.Cog, name="Miku"):
 
         if query.startswith(("http://", "https://")):
             song = await Song.from_url_and_download(query, author)
+
             if song:
                 await state.queue.put(song)
                 response_message = f"✅ Đã thêm **{song.title}** vào hàng đợi."
+
                 if isinstance(ctx, discord.Interaction) and ctx.response.is_done():
                     await ctx.followup.send(response_message)
                 else:
                     await self._send_response(ctx, response_message)
+
                 if state.player_task is None or state.player_task.done():
                     state.player_task = asyncio.create_task(state.player_loop())
             else:
@@ -907,10 +351,12 @@ class MainCog(commands.Cog, name="Miku"):
 
     async def _seek_logic(self, ctx: AnyContext, timestamp: str):
         state = self.get_guild_state(ctx.guild.id)
+
         if not state.voice_client or not state.current_song:
             return await self._send_response(
                 ctx, "Không có bài hát nào đang phát để tua.", ephemeral=True
             )
+
         match = re.match(r"(?:(\d+):)?(\d+)", timestamp)
         if not match:
             try:
@@ -925,24 +371,28 @@ class MainCog(commands.Cog, name="Miku"):
             minutes = int(match.group(1) or 0)
             seconds = int(match.group(2))
             seconds += minutes * 60
+
         if not 0 <= seconds < state.current_song.duration:
             return await self._send_response(
                 ctx, "Không thể tua đến thời điểm không hợp lệ.", ephemeral=True
             )
-        state.is_seeking = True
-        ffmpeg_options_seek = FFMPEG_OPTIONS.copy()
-        ffmpeg_options_seek["before_options"] = f"-ss {seconds}"
-        new_source = discord.PCMVolumeTransformer(
-            discord.FFmpegPCMAudio(state.current_song.filepath, **ffmpeg_options_seek),
-            volume=state.volume,
-        )
-        state.voice_client.stop()
-        state.voice_client.play(
-            new_source,
-            after=lambda e: self.bot.loop.call_soon_threadsafe(
-                state.song_finished_event.set
-            ),
-        )
+
+        # state.is_seeking = True
+        # ffmpeg_options_seek = FFMPEG_OPTIONS.copy()
+        # ffmpeg_options_seek["before_options"] = f"-ss {seconds}"
+        # new_source = discord.PCMVolumeTransformer(
+        #     discord.FFmpegPCMAudio(state.current_song.filepath, **ffmpeg_options_seek),
+        #     volume=state.volume,
+        # )
+
+        # state.voice_client.stop()
+        # state.voice_client.play(
+        #     new_source,
+        #     after=lambda e: self.bot.loop.call_soon_threadsafe(
+        #         state.song_finished_event.set
+        #     ),
+        # )
+
         await self._send_response(ctx, f"⏩ Đã tua đến `{seconds}` giây.")
 
     async def _shuffle_logic(self, ctx: AnyContext):
@@ -953,10 +403,13 @@ class MainCog(commands.Cog, name="Miku"):
             )
         queue_list = list(state.queue._queue)
         random.shuffle(queue_list)
+
         while not state.queue.empty():
             state.queue.get_nowait()
+
         for song in queue_list:
             await state.queue.put(song)
+
         await self._send_response(ctx, "🔀 Đã xáo trộn hàng đợi!")
 
     async def _remove_logic(self, ctx: AnyContext, index: int):
@@ -965,13 +418,16 @@ class MainCog(commands.Cog, name="Miku"):
             return await self._send_response(
                 ctx, "Số thứ tự không hợp lệ.", ephemeral=True
             )
+
         queue_list = list(state.queue._queue)
         removed_song = queue_list.pop(index - 1)
         removed_song.cleanup()
+
         while not state.queue.empty():
             state.queue.get_nowait()
         for song in queue_list:
             await state.queue.put(song)
+
         await self._send_response(
             ctx, f"🗑️ Đã xóa **{removed_song.title}** khỏi hàng đợi."
         )
@@ -1153,8 +609,7 @@ class MainCog(commands.Cog, name="Miku"):
     async def slash_lyrics(self, interaction: discord.Interaction):
         await self._lyrics_logic(interaction)
 
-
 async def setup(bot: commands.Bot):
     """Thiết lập và đăng ký các cogs vào bot."""
-    await bot.add_cog(MainCog(bot))
-    log.info("Đã thêm cog chính (MainCog).")
+    await bot.add_cog(MusicCog(bot))
+    log.info("Đã thêm cog chính (MusicCog).")
