+++ enums/LoopMode.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class LoopMode(Enum):
+    OFF = 0
+    SONG = 1
+    QUEUE = 2
